requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
astroid==2.15.6
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
dataclasses-json==0.5.14
dill==0.3.7
Flask==2.3.2
frozenlist==1.4.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.263
langsmith==0.0.22
lazy-object-proxy==1.9.0
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
platformdirs==3.10.0
pycparser==2.21
pydantic==1.10.12
pylint==2.17.5
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
tenacity==8.2.2
tomlkit==0.12.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
validators==0.21.0
weaviate-client==3.22.1
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.2
